Forgot to add commit message.
Okay, so the last commit was a major update by Tcll which rewrote a huge portion of the code, changed the menu, and added something to remove the need for static versions. This is awesome and I hope the installer becomes more usable and friendly.
It's a bit unreadable in exchange for performance, since that's Tclls workflow, but I think overall it is a major improvement.
Next update may scrap this and replace it with a bigger project.