Make updater update in the directory that the script is in
- Changed update filename to be more readable
- Add filepath variable which is the directory that the script is in
- Change LinuxED updater to work with filepath

In a previous commit I changed dirpath to be the path relative to where the shell is, which essentially allows you to use LinuxED as a system-wide script. This commit fixes the updater to make it completely functional system-wide.
I find this to be a much better alternative to $XDG_DATA_HOME as this allows the user to specify where they want their EnhancedDiscord installation while keeping updating functional.